integrated styles into the rest of the app
