+++ app/src/main/java/com/example/bluetoothchat/MainSetting.java
